+++ test/utils/XhrUtil.spec.js
@@ -25,5 +25,72 @@ describe('XhrUtil', () => {
       const promise = XhrUtil.send({ url: 'https://vwo.com', method: 'GET' }).catch(() => {});
       expect(typeof promise.then).toBe('function');
     });
+
+    describe('method: _getStoredSettings', () => {
+      it('should return false and no settings if userStorageService is not passed', () => {
+        const data = XhrUtil._getStoredSettings();
+
+        expect(data.isStoredData).toBe(false);
+        expect(data.parsedSettings).toBe(undefined);
+      });
+
+      it('should return false and no settings if userStorageService is not valid', () => {
+        const userStorageService = {};
+        const data = XhrUtil._getStoredSettings(userStorageService);
+
+        expect(data.isStoredData).toBe(false);
+        expect(data.parsedSettings).toBe(undefined);
+      });
+
+      it('should return false and no settings if userStorageService returns blank object', () => {
+        const userStorageService = {
+          getSettings: function() {
+            return {};
+          }
+        };
+        const data = XhrUtil._getStoredSettings(userStorageService);
+
+        expect(data.isStoredData).toBe(false);
+        expect(data.parsedSettings).toEqual(undefined);
+      });
+
+      it('should return false and no settings if userStorageService returns stringified blank object', () => {
+        const userStorageService = {
+          getSettings: function() {
+            return JSON.stringify({});
+          }
+        };
+        const data = XhrUtil._getStoredSettings(userStorageService);
+
+        expect(data.isStoredData).toBe(false);
+        expect(data.parsedSettings).toEqual({});
+      });
+
+      it('should return false and settings if userStorageService returns invalid stringifeid data', () => {
+        const settings = { accountId: 123, faultyKeys: [] };
+        const userStorageService = {
+          getSettings: function() {
+            return JSON.stringify(settings);
+          }
+        };
+        const data = XhrUtil._getStoredSettings(userStorageService);
+
+        expect(data.isStoredData).toBe(false);
+        expect(data.parsedSettings).toEqual(settings);
+      });
+
+      it('should return true and settings if userStorageService returns valid settings', () => {
+        const settings = { accountId: 123, campaigns: [], version: 1, sdkKey: 'abc' };
+        const userStorageService = {
+          getSettings: function() {
+            return JSON.stringify(settings);
+          }
+        };
+        const data = XhrUtil._getStoredSettings(userStorageService);
+
+        expect(data.isStoredData).toBe(true);
+        expect(data.parsedSettings).toEqual(settings);
+      });
+    });
   });
 });